+++ api/server/middleware/abortControllers.js
@@ -1,0 +1,2 @@
+// abortControllers.js
+module.exports = new Map();
