+++ api/server/middleware/setHeaders.js
@@ -1,0 +1,12 @@
+function setHeaders(req, res, next) {
+  res.writeHead(200, {
+    Connection: 'keep-alive',
+    'Content-Type': 'text/event-stream',
+    'Cache-Control': 'no-cache, no-transform',
+    'Access-Control-Allow-Origin': '*',
+    'X-Accel-Buffering': 'no',
+  });
+  next();
+}
+
+module.exports = setHeaders;
