+++ api/server/routes/ask/openAI.js
@@ -1,74 +1,33 @@
 const express = require('express');
 const router = express.Router();
-const { titleConvo, OpenAIClient } = require('../../../app');
-const { parseConvo } = require('librechat-data-provider');
-const { getAzureCredentials, abortMessage } = require('../../../utils');
+const { titleConvo } = require('../../../app');
+const { buildOptions, initializeClient } = require('../openAI');
+const { sendMessage, createOnProgress } = require('../../utils');
 const { saveMessage, getConvoTitle, saveConvo, getConvo } = require('../../../models');
-const { handleError, sendMessage, createOnProgress } = require('../handlers');
-const requireJwtAuth = require('../../middleware/requireJwtAuth');
+const {
+  handleAbort,
+  createAbortController,
+  handleAbortError,
+  setHeaders,
+  requireJwtAuth,
+} = require('../../middleware');
 
-const abortControllers = new Map();
+router.post('/abort', requireJwtAuth, handleAbort());
 
-router.post('/abort', requireJwtAuth, async (req, res) => {
-  try {
-    return await abortMessage(req, res, abortControllers);
-  } catch (err) {
-    console.error(err);
-  }
-});
-
-router.post('/', requireJwtAuth, async (req, res) => {
-  const { endpoint, text, parentMessageId, conversationId } = req.body;
-  if (text.length === 0) {
-    return handleError(res, { text: 'Prompt empty or too short' });
-  }
-  const isOpenAI = endpoint === 'openAI' || endpoint === 'azureOpenAI';
-  if (!isOpenAI) {
-    return handleError(res, { text: 'Illegal request' });
-  }
-
-  // build endpoint option
-  const parsedBody = parseConvo(endpoint, req.body);
-  const { chatGptLabel, promptPrefix, ...rest } = parsedBody;
-  const endpointOption = {
-    chatGptLabel,
-    promptPrefix,
-    modelOptions: {
-      ...rest,
-    },
-  };
-
-  console.log('ask log');
-  console.dir({ text, conversationId, endpointOption }, { depth: null });
-
+router.post('/', requireJwtAuth, setHeaders, async (req, res) => {
+  const { text, endpointOption, conversationId, parentMessageId } = buildOptions(req, res);
   // eslint-disable-next-line no-use-before-define
   return await ask({
     text,
     endpointOption,
     conversationId,
     parentMessageId,
-    endpoint,
     req,
     res,
   });
 });
 
-const ask = async ({
-  text,
-  endpointOption,
-  parentMessageId = null,
-  endpoint,
-  conversationId,
-  req,
-  res,
-}) => {
-  res.writeHead(200, {
-    Connection: 'keep-alive',
-    'Content-Type': 'text/event-stream',
-    'Cache-Control': 'no-cache, no-transform',
-    'Access-Control-Allow-Origin': '*',
-    'X-Accel-Buffering': 'no',
-  });
+const ask = async ({ text, endpointOption, parentMessageId = null, conversationId, req, res }) => {
   let metadata;
   let userMessage;
   let userMessageId;
@@ -112,7 +71,7 @@ const ask = async ({
     },
   });
 
-  const abortController = new AbortController();
+  const { abortController, onStart } = createAbortController(res, req, endpointOption);
   abortController.abortAsk = async function () {
     this.abort();
 
@@ -139,29 +98,8 @@ const ask = async ({
     };
   };
 
-  const onStart = (userMessage) => {
-    sendMessage(res, { message: userMessage, created: true });
-    abortControllers.set(userMessage.conversationId, { abortController, ...endpointOption });
-  };
-
   try {
-    const clientOptions = {
-      // debug: true,
-      // contextStrategy: 'refine',
-      reverseProxyUrl: process.env.OPENAI_REVERSE_PROXY || null,
-      proxy: process.env.PROXY || null,
-      endpoint,
-      ...endpointOption,
-    };
-
-    let openAIApiKey = req.body?.token ?? process.env.OPENAI_API_KEY;
-
-    if (process.env.AZURE_API_KEY && endpoint === 'azureOpenAI') {
-      clientOptions.azure = JSON.parse(req.body?.token) ?? getAzureCredentials();
-      openAIApiKey = clientOptions.azure.azureOpenAIApiKey;
-    }
-
-    const client = new OpenAIClient(openAIApiKey, clientOptions);
+    const { client, openAIApiKey } = initializeClient(req, endpointOption);
 
     let response = await client.sendMessage(text, {
       user,
@@ -208,7 +146,7 @@ const ask = async ({
         text,
         response,
         openAIApiKey,
-        azure: endpoint === 'azureOpenAI',
+        azure: endpointOption.endpoint === 'azureOpenAI',
       });
       await saveConvo(req.user.id, {
         conversationId,
@@ -216,24 +154,14 @@ const ask = async ({
       });
     }
   } catch (error) {
-    console.error(error);
     const partialText = getPartialText();
-    if (partialText?.length > 2) {
-      return await abortMessage(req, res, abortControllers);
-    } else {
-      const errorMessage = {
-        messageId: responseMessageId,
-        sender: 'ChatGPT',
-        conversationId,
-        parentMessageId: userMessageId,
-        unfinished: false,
-        cancelled: false,
-        error: true,
-        text: error.message,
-      };
-      await saveMessage(errorMessage);
-      handleError(res, errorMessage);
-    }
+    handleAbortError(res, req, error, {
+      partialText,
+      conversationId,
+      sender: 'ChatGPT',
+      messageId: responseMessageId,
+      parentMessageId: userMessageId,
+    });
   }
 };
