+++ api/utils/index.js
@@ -1,16 +1,10 @@
 const azureUtils = require('./azureUtils');
-const cryptoUtils = require('./crypto');
 const { tiktokenModels, maxTokensMap } = require('./tokens');
-const sendEmail = require('./sendEmail');
-const abortMessage = require('./abortMessage');
 const findMessageContent = require('./findMessageContent');
 
 module.exports = {
-  ...cryptoUtils,
   ...azureUtils,
   maxTokensMap,
   tiktokenModels,
-  sendEmail,
-  abortMessage,
   findMessageContent,
 };
